% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/222_export_opt.r
\name{Initial Control Functions}
\alias{.rpcopt_init}
\alias{Initial Control Functions}
\title{Initial controls in pbdRPC}
\usage{
.rpcopt_init(envir = .GlobalEnv)
}
\arguments{
\item{envir}{an environment where RPC controls locate}
}
\value{
\code{.rpcopt_init()} initial the RPC control
at \code{envir}.
}
\description{
Initial control functions
}
\details{
\code{.rpcopt_init()} initials default RPC controls.
}
\examples{
\dontrun{
library(pbdRPC, quietly = TRUE)
rpcopt_set(user = "snoweye", hostname = "192.168.56.101")

ls(.pbd_env)
rm(.pbd_env)
.rpcopt_init()
ls(.pbd_env)

.pbd_env$RPC.LI
}

}
\author{
Wei-Chen Chen \email{wccsnow@gmail.com}.
}
\references{
ZeroMQ/4.1.0 API Reference:
\url{http://api.zeromq.org/4-1:_start}

Programming with Big Data in R Website: \url{http://r-pbd.org/}
}
\seealso{
\code{.pbd_env}.
}
\keyword{programming}

