% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_plink.r
\name{find_plink}
\alias{find_plink}
\title{Find plink}
\usage{
find_plink()
}
\value{
A full path a full path to the \code{plink} or \code{plink.exe}
command if found. Otherwise, "plink" is returned.
}
\description{
Find the plink internal compiled with pbdRPC package.
}
\details{
The function returns a full path to the \code{plink} or \code{plink.exe}
command if found.
}
\examples{
\dontrun{
library(pbdRPC, quietly = TRUE)

find_plink()
}

}

