% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine.r
\name{print-machine}
\alias{print-machine}
\alias{print.machine}
\title{print-machine}
\usage{
\method{print}{machine}(x, ...)
}
\arguments{
\item{x}{machine class object}

\item{...}{ignored}
}
\description{
Printing for machine class objects.
}
