% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpc_pid.r
\name{rpc_pid}
\alias{rpc_kill}
\alias{rpc_pid}
\alias{rpc_ps}
\title{Search or Kill pid of Remote Procedure Call}
\usage{
rpc_ps()

rpc_kill(pid)
}
\arguments{
\item{pid}{a vector containing proccess id's in interger.}
}
\value{
List or kill all possible pid's.
}
\description{
Search or kill possible pid's that running remote procedure calls
in background.
}
\details{
\code{rpc_ps()} prints possible pid's that running remote procedure
calls in background via one of
\code{ssh}, \code{plink}, \code{plink.exe}, or \code{cmd.exe}.

\code{rpc_kill()} kill pid's.
}

