% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/222_export_env.r
\docType{data}
\name{RPC Control Environment}
\alias{.pbd_env}
\alias{RPC Control Environment}
\title{Sets of controls in pbdRPC.}
\format{Objects contain several parameters for communicators and methods.}
\description{
These sets of controls are used to provide default values in this package.
}
\details{
The elements of \code{.pbd_env$RPC.CT} are default values for looking RPC
control files.

The elements of \code{.pbd_env$RPC.LI} are default values for RPC defult
login information.
}
\author{
Wei-Chen Chen \email{wccsnow@gmail.com}.

Programming with Big Data in R Website: \url{http://r-pbd.org/}
}
\seealso{
\code{\link{.rpcopt_init}()}.
}
\keyword{global}
\keyword{variables}

