% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbdRPC-package.r
\docType{package}
\name{pbdRPC-package}
\alias{pbdRPC-package}
\title{pbdRPC}
\description{
A very light yet secure implementation of remote procedure call.
}
\references{
Project URL: \url{https://github.com/RBigData/pbdRPC}
}
\author{
Wei-Chen Chen and Drew Schmidt
}
\keyword{package}
